+++ src/settings.ts
@@ -494,6 +494,18 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
 
         if (plugin.gitManager instanceof SimpleGit)
             new Setting(containerEl)
+                .setName("Additional environment variables")
+                .setDesc("Use each line for a new environment variable in the format KEY=VALUE")
+                .addTextArea((cb) => {
+                    cb.setPlaceholder("GIT_DIR=/path/to/git/dir");
+                    cb.setValue(plugin.localStorage.getEnvVars().join("\n"));
+                    cb.onChange((value) => {
+                        plugin.localStorage.setEnvVars(value.split("\n"));
+                    });
+                });
+
+        if (plugin.gitManager instanceof SimpleGit)
+            new Setting(containerEl)
                 .setName("Additional PATH environment variable paths")
                 .setDesc("Use each line for one path")
                 .addTextArea((cb) => {
@@ -504,7 +516,8 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
                 });
         if (plugin.gitManager instanceof SimpleGit)
             new Setting(containerEl)
-                .setName("Reload with new PATH environment variable")
+                .setName("Reload with new environment variables")
+                .setDesc("Removing previously added environment variables will not take effect until Obsidian is restarted.")
                 .addButton(cb => {
                     cb.setButtonText("Reload");
                     cb.setCta();
@@ -527,6 +540,19 @@ export class ObsidianGitSettingsTab extends PluginSettingTab {
                     plugin.settings.basePath = value;
                     plugin.saveSettings();
                     plugin.gitManager.updateBasePath(value || "");
+                });
+            });
+
+        new Setting(containerEl)
+            .setName("Custom Git directory path (Instead of '.git')")
+            .setDesc(`Requires restart of Obsidian to take effect. Use "\\" instead of "/" on Windows.`)
+            .addText((cb) => {
+                cb.setValue(plugin.settings.gitDir);
+                cb.setPlaceholder(".git");
+                cb.onChange((value) => {
+                    plugin.settings.gitDir = value;
+                    plugin.saveSettings();
+
                 });
             });